Make config module less modularized
When compiling boost with clang we get errors from the config dir.
The reason here is that this directory is doing a lot of preprocessor
magic and that doesn't work well across submodules. let's just make
the top-level config header a module so that the preprocessor magic
is contained in a single TU (which fixes the compilation of the
'config' module). Right now it's failing with this on the latest
clang trunk:

While building module 'config' imported from /home/teemperor/boost-compile/inc/boost/type_traits/detail/config.hpp:13:
In file included from <module-includes>:34:
/home/teemperor/boost-compile/inc/boost/config/requires_threads.hpp:47:5: error: "Compiler threading support is not turned on. Please set the correct command line options for threading:
      -pthread (Linux), -pthreads (Solaris) or -mthreads (Mingw32)"
    ^